Unify light/dark tray icon switching
Before this change, the app displayed a dark tray icon on Linux and a
red one on Windows, regardless of the OS's dark mode setting. On macOS,
it used the OS mechanism via tray image names *Template.png, allowing
macOS to automatically invert the tray icon colors.

With this change, the app now relies on the Electron API: nativeTheme.
This works on macOS, Linux, and Windows. When the OS 'Dark Mode' setting
is enabled, the app will choose a white tray icon; otherwise, it will
use a black one.

Additionally, there is a new setting that allows the user to force the
tray icon to be either black or white. This is useful in situations
where a light theme is in use, but the tray panel is still black, which
is the default configuration on Ubuntu, for example.